+++ src/dataset.js
@@ -636,16 +636,19 @@ const getOrCreateDataset = (datasetIdOrName) => {
  * @param {string} [datasetIdOrName]
  *   ID or name of the dataset to be opened. If `null` or `undefined`,
  *   the function returns the default dataset associated with the actor run.
+ * @param {boolean} [enforceCloud=false]
+ *   If set to `true` then enforces cloud storage usage. This way is possible to combine local and
+ *   cloud storages when devloping locally.
  * @returns {Promise<Dataset>}
  * @memberof module:Apify
  * @name openDataset
  * @function
  */
-export const openDataset = (datasetIdOrName) => {
+export const openDataset = (datasetIdOrName, enforceCloud = false) => {
     checkParamOrThrow(datasetIdOrName, 'datasetIdOrName', 'Maybe String');
     ensureTokenOrLocalStorageEnvExists('dataset');
 
-    return process.env[ENV_VARS.LOCAL_STORAGE_DIR]
+    return process.env[ENV_VARS.LOCAL_STORAGE_DIR] && !enforceCloud
         ? openLocalStorage(datasetIdOrName, ENV_VARS.DEFAULT_DATASET_ID, DatasetLocal, datasetsCache)
         : openRemoteStorage(datasetIdOrName, ENV_VARS.DEFAULT_DATASET_ID, Dataset, datasetsCache, getOrCreateDataset);
 };