+++ src/key_value_store.js
@@ -454,16 +454,19 @@ const getOrCreateKeyValueStore = (storeIdOrName) => {
  * @param {string} [storeIdOrName]
  *   ID or name of the key-value store to be opened. If `null` or `undefined`,
  *   the function returns the default key-value store associated with the actor run.
+ * @param {boolean} [enforceCloud=false]
+ *   If set to `true` then enforces cloud storage usage. This way is possible to combine local and
+ *   cloud storages when devloping locally.
  * @returns {Promise<KeyValueStore>}
  * @memberof module:Apify
  * @name openKeyValueStore
  * @function
  */
-export const openKeyValueStore = (storeIdOrName) => {
+export const openKeyValueStore = (storeIdOrName, enforceCloud = false) => {
     checkParamOrThrow(storeIdOrName, 'storeIdOrName', 'Maybe String');
     ensureTokenOrLocalStorageEnvExists('key value store');
 
-    return process.env[ENV_VARS.LOCAL_STORAGE_DIR]
+    return process.env[ENV_VARS.LOCAL_STORAGE_DIR] && !enforceCloud
         ? openLocalStorage(storeIdOrName, ENV_VARS.DEFAULT_KEY_VALUE_STORE_ID, KeyValueStoreLocal, storesCache)
         : openRemoteStorage(storeIdOrName, ENV_VARS.DEFAULT_KEY_VALUE_STORE_ID, KeyValueStore, storesCache, getOrCreateKeyValueStore);
 };