+++ test/request_queue.js
@@ -548,6 +548,11 @@ describe('RequestQueue', () => {
             mock.expects('openLocalStorage').once();
             Apify.openRequestQueue();
 
+            mock.expects('openLocalStorage').once();
+            Apify.openRequestQueue('xxx');
+            mock.expects('openRemoteStorage').once();
+            Apify.openRequestQueue('xxx', true);
+
             delete process.env[ENV_VARS.LOCAL_STORAGE_DIR];
             process.env[ENV_VARS.TOKEN] = 'xxx';